Rename to arbitrage
