+++ src/index.ts
@@ -8,7 +8,7 @@ import { handler as vestedBalanceHandler } from "./handlers/vestedBalance";
 import { handler as delegatorCount } from "./handlers/delegatorCount";
 import { handler as totalDelegators } from "./handlers/totalDelegators";
 import { handler as totalStakedCoins } from "./handlers/totalStakedCoins";
-import { handler as coinPriceHandler } from "./handlers/coinPrice";
+import { handler as coinPriceHandler } from "./handlers/flagArbitrage";
 import { webhookTriggers } from "./handlers/webhookTriggers";
 
 addEventListener("scheduled", (event: any) => {
